# src/main/java/org/example/datahub/api/FilesApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.1.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package org.example.datahub.api;

import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2025-12-04T09:26:02.924376840+08:00[Asia/Shanghai]")
@Validated
@Tag(name = "files", description = "the files API")
public interface FilesApi {

    /**
     * GET /files/{file_id} : Get file
     * Get file by file id.
     *
     * @param fileId  (required)
     * @return File response (status code 200)
     */
    @Operation(
        operationId = "getFile",
        summary = "Get file",
        description = "Get file by file id.",
        responses = {
            @ApiResponse(responseCode = "200", description = "File response", content = {
                @Content(mediaType = "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet", schema = @Schema(implementation = org.springframework.core.io.Resource.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/files/{file_id}",
        produces = { "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet" }
    )
    
    ResponseEntity<org.springframework.core.io.Resource> getFile(
        @Parameter(name = "file_id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("file_id") Long fileId
    );

}