# src/main/resources/openapi/api.yml
openapi: 3.0.3
info:
  title: Research Collection System API
  version: 1.0.0
  description: API specification for research task collection system.
servers:
  - url: /api
paths:
  /auth/login:
    post:
      operationId: userLogin
      summary: User login
      description: Authenticate user and return JWT.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserLoginRequest"
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserLoginResponse"

#  /auth/logout:
#    post:
#      operationId: userLogout
#      summary: User logout
#      description: Sign out user.
#      responses:
#        "200":
#          description: Successful response
#          content:
#            application/json:
#              schema:
#                $ref: "#/components/schemas/SuccessResponse"
...
  /files/{file_id}:
    get:
      operationId: getFile
      summary: Get file
      description: Get file by file id.
      parameters:
        - in: path
          name: file_id
          required: true
          schema: { type: integer, format: int64 }
      responses:
        "200":
          description: File response
          content:
            application/vnd.openxmlformats-officedocument.spreadsheetml.sheet:
              schema:
                type: string
                format: binary
components:
  schemas:
...
    UserLoginRequest:
      type: object
      properties:
        username: { type: string }
        password: { type: string }
      required: [username, password]
    UserLoginResponse:
      type: object
      properties:
        success: { type: boolean }
        data:
          type: object
          properties:
            token: { type: string }
            user:
              type: object
              properties:
                user_id: { type: integer, format: int64 }
                username: { type: string }
                role: { type: string }
        message: { type: string }

    # ==================== USER ====================
...
    SuccessResponse:
      type: object
      properties:
        success: { type: boolean }
        message: { type: string }
...